+++ test/fixtures/duplicate-exports-project/src/index.ts
@@ -1,0 +1,2 @@
+export { isUntagged, isUntaggedAlias, isAlias, reExportedAlias } from "./helpers";
+export { default as defaultHandler } from "./default-handler";